app/python/requirements.txt
--extra-index-url https://PySimpleGUI.net/install
PySimpleGUI
psutil==5.9.8
ttkbootstrap
PyYAML==6.0.2
